Implements trust requirement banners
