+++ app/dashboard/templates/profiles/profile.html
@@ -149,6 +149,8 @@
         {% endif %}
       </script>
       {% bundle merge_js file profile %}
+        <script src="{% static "v2/js/parseMonthDay.js" %}"></script>
+        <script src="{% static "v2/js/why-this-matters-modal.js" %}"></script>
         <script src="{% static "v2/js/pages/tabs.js" %}"></script>
         <script src="{% static "v2/js/pages/profile.js" %}"></script>
         <script src="{% static "v2/js/pages/tribe_title.js" %}"></script>
