+++ app/grants/templates/grants/shared/landing_qf_active.html
@@ -1,5 +1,34 @@
 {% load humanize static i18n tz %}
 <div class="container bg-white">
+
+  <!-- Trust banner -->
+
+  <div v-if="{{trust_bonus}} < 150" v-cloak class="alert alert-primary spacer-mt-4 border-1 border-primary">
+    <i :class="'far font-weight-bold mr-4 mt-1 position-absolute cursor-pointer' + (!bannerCollapsed ? ' fa-minus' : ' fa-plus')" @click="toggleBannerCollapse" style="right: 0px; z-index: 2"></i>
+    <div class="row">
+      <div :class="(!bannerCollapsed ? 'col-12 col-md-auto text-center pb-3' : 'col-auto text-center my-auto my-md-0')" style="max-width: 90px;">
+        <object type="image/svg+xml" data="{% static 'v2/images/trust/64x64-green.svg' %}" height="65" :height="(!bannerCollapsed ? '65' : '35')"></object>
+      </div>
+      <div :class="(!bannerCollapsed ? 'col-12 col-md-8 col-lg-7 col-xl-9 mt-1' : 'col-8 col-lg-7 col-xl-9 mt-1')">
+        <h6 class="gc-font-base mt-1">Maximize Your Impact! Your current match is {{ trust_bonus }}%</h6>
+        <span v-if="!bannerCollapsed" class="font-smaller-1 line-height-2">Verified users have up to 150% of their donation amount count towards the match fund algorithm. Unverified users — only 50%. Maximize your match by verifying with identity partners.</span>
+        <div v-if="!bannerCollapsed" class="row spacer-mt-2 font-smaller-1 line-height-2">
+          <div class="col-12 col-lg-6 spacer-mt-2">
+            {% timezone "US/Mountain" %}
+              <a><i class="far fa-calendar-alt spacer-pr-2 text-primary"></i> Verify during {{ round_start_date|localtime|date:"F j" }} - {{ round_end_date|localtime|date:"F j" }}</a>
+            {% endtimezone %}
+          </div>
+          <div class="col-12 col-lg-6 spacer-mt-2">
+            <span @click="showModal('why-this-matters-modal')" class="cursor-pointer hover-underline"><i class="fas fa-play-circle spacer-pr-2 text-primary"></i> Why this matters</span>
+          </div>
+        </div>
+      </div>
+      <div v-if="!bannerCollapsed" class="col-md-12 col-lg-3 col-xl-2 mt-4 m-lg-auto">
+        <a href="/profile/trust" class="btn btn-primary px-5 py-2 w-100 w-lg-auto">Verify</a>
+      </div>
+    </div>
+  </div>
+
   <!--  Header section -->
   <div class="row justify-content-between align-items-center my-5">
     <div class="col">
@@ -86,16 +115,20 @@
     </div> -->
   </section>
 
-   <!-- Ongoing matching pools (CURRENTLY HIDDEN AND NOT IMPLEMENTED) -->
-   {% comment %}
-   <div class="my-5">
-     <div class="row justify-content-start align-items-center">
-       <div class="col-auto"><img height="30rem" src="{% static 'v2/images/ethereum.svg' %}" /></div>
-       <h4 class="col-auto font-weight-bold py-0 my-0 pl-0">Ongoing Matching Pools</h4>
-     </div>
-   </div>
-   {% endcomment %}
+  <!-- Ongoing matching pools (CURRENTLY HIDDEN AND NOT IMPLEMENTED) -->
+  {% comment %}
+    <div class="my-5">
+      <div class="row justify-content-start align-items-center">
+        <div class="col-auto"><img height="30rem" src="{% static 'v2/images/ethereum.svg' %}" /></div>
+        <h4 class="col-auto font-weight-bold py-0 my-0 pl-0">Ongoing Matching Pools</h4>
+      </div>
+    </div>
+  {% endcomment %}
 
 
-  </div>
+  <why-this-matters-modal
+    :show-modal="visibleModal === 'why-this-matters-modal'"
+    @modal-dismissed="hideModal()">
+  </why-this-matters-modal>
+</div>
